Add files via upload
Python program that that defines an Account class with ID, balance, and Annual Interest Rate variables. Define methods to perform functions such as withdraw, deposit, check balance as well as set and get the annual and monthly interest rate. Includes an ATM function that creates 10 instances of the account and uses an infinite loop to perform the methods for an account.